(PC-17810) refactor(bookOffer): add context folder
