+++ src/features/bookOffer/components/BookDuoChoice.web.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
+import { BookingState, Step } from 'features/bookOffer/context/reducer'
 import { mockOffer } from 'features/bookOffer/fixtures/offer'
-import { BookingState, Step } from 'features/bookOffer/pages/reducer'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { fireEvent, render } from 'tests/utils/web'
