+++ src/features/bookOffer/components/BookingEventChoices.tsx
@@ -7,8 +7,8 @@ import { BookDateChoice } from 'features/bookOffer/components/BookDateChoice'
 import { BookDuoChoice } from 'features/bookOffer/components/BookDuoChoice'
 import { BookHourChoice } from 'features/bookOffer/components/BookHourChoice'
 import { BookingDetails } from 'features/bookOffer/components/BookingDetails'
+import { Step } from 'features/bookOffer/context/reducer'
 import { useBooking } from 'features/bookOffer/helpers/useBooking'
-import { Step } from 'features/bookOffer/pages/reducer'
 import { useCreditForOffer } from 'features/offer/helpers/useHasEnoughCredit/useHasEnoughCredit'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { getSpacing, Spacer } from 'ui/theme'
