+++ src/features/bookOffer/helpers/useModalContent.tsx
@@ -6,9 +6,9 @@ import { AlreadyBooked } from 'features/bookOffer/components/AlreadyBooked'
 import { BookingDetails } from 'features/bookOffer/components/BookingDetails'
 import { BookingEventChoices } from 'features/bookOffer/components/BookingEventChoices'
 import { BookingImpossible } from 'features/bookOffer/components/BookingImpossible'
+import { Step } from 'features/bookOffer/context/reducer'
 import { useBooking } from 'features/bookOffer/helpers/useBooking'
 import { useBookingOffer } from 'features/bookOffer/helpers/useBookingOffer'
-import { Step } from 'features/bookOffer/pages/reducer'
 import { getOfferPrice } from 'features/offer/helpers/getOfferPrice/getOfferPrice'
 import { useSubcategoriesMapping } from 'libs/subcategories'
 import { ModalLeftIconProps } from 'ui/components/modals/types'
