+++ src/features/bookOffer/pages/BookingContext.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { initialBookingState } from 'features/bookOffer/pages/reducer'
+import { initialBookingState } from 'features/bookOffer/context/reducer'
 import { IBookingContext } from 'features/bookOffer/types'
 
 export const BookingContext = React.createContext<IBookingContext>({
